requirements.txt
appdirs==1.4.4
asgiref==3.5.2
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
cmake==3.22.4
colorama==0.4.4
cycler==0.10.0
Cython==0.29.21
defusedxml==0.7.1
Django==3.1.8
django-allauth==0.42.0
django-ckeditor==6.3.2
django-crispy-forms==1.10.0
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-pandas==0.6.2
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
imutils==0.5.0
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.3.3
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
regex==2022.4.24
requests==2.25.1
requests-oauthlib==1.3.1
scikit-learn==1.1.0
scipy==1.8.1
seaborn==0.11.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
typed-ast==1.5.3
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9